Minor bugs have been fixed, and the possibility of a client creating a user with a nickname already present in the database has been addressed. This is checked during user creation in the create method. If a user with such an ID is found, a runtime exception is thrown and handled in the AdapterUserImpl class. Additionally, the return type of the registration method in the AdapterUser interface has been changed from void to boolean. The data type of the Note date has also been changed from string to long.

+++ src/main/java/org/mycorp/adapters/AdapterNoteImpl.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Component;
 
 @Component
 public class AdapterNoteImpl implements AdapterNote{
-
     final UserCategoryLinkService userCategoryLinkService;
     final NoteService noteService;
 
@@ -23,7 +22,7 @@ public class AdapterNoteImpl implements AdapterNote{
     public boolean createNote(int id_user, int id_category, Note note) {
         UserCategoryLink link = userCategoryLinkService.findLink(id_user, id_category);
 
-        if(link==null)
+        if (link == null)
             return false;
 
         note.setLink(link);
@@ -43,7 +42,12 @@ public class AdapterNoteImpl implements AdapterNote{
     }
 
     @Override
-    public boolean deleteNote(int id) {
+    public boolean deleteNote(int id_user, int id_category, int id) {
+        UserCategoryLink link = userCategoryLinkService.findLink(id_user, id_category);
+
+        if(link==null)
+            return false;
+
         return noteService.delete(id);
     }
 }
